cast away a few warnings
--HG--
extra : convert_revision : svn%3A1766ff46-f334-0410-ab20-d63176f87757/trunk%40810

+++ apps/lascommon.c
@@ -445,7 +445,7 @@ void print_header(LASHeaderH header, const char* file_name) {
     fprintf(stderr, "  VLR Summary\n");
     fprintf(stderr, "---------------------------------------------------------\n");
 
-        for (i = 0; i < nVLR; i++) {
+        for (i = 0; i < (int)nVLR; i++) {
             pVLR = LASHeader_GetVLR(header, i);
 
             if (LASError_GetLastErrorNum()) {
